Ignore pong responses from websocket_client
When `:websocket_client` is started with `keepalive: timeout`, it
actively pings the server using the standard Websocket mechanism. It
reports responses as `:pong`. Since it handles all of the timers,
`phoenix_client` doesn't need to do anything.

This commit ignores the `:pong` responses since they fill the log with
lines like this:

```
13:32:23.941 [warn]  Unknown message {:pong, "foo"}
```