Move heartbeat functionality to transport
The current heartbeat implementation sends messages to the server to
keep the connection up, but it doesn't close the connection if a pong
responses isn't received. This means that if the server stops responding
then this won't be detected. On Linux, the system's TCP keepalive will
eventually detect this condition, but that's after 2 hours.

Rather than fixing the heartbeat implementation here, delegate it to the
transport. For the websocket case, `websocket_client` has a keepalive
implementation already. It sends payload-less websocket ping messages.
Since websockets are terminated on the same server as the Phoenix
channel, there's no loss in what failures can be detected.

In theory, Phoenix channels are transport agnostic, so there is the
possibility that delegating the keepalive to the transport either won't
work or won't detect connection loss end-to-end. If such a transport is
ever created and supported, it may be necessary to implement the
heartbeat code for it.