Make websocket_client a required dependency
Since websocket_client is the only supported transport and there's not
an alternative like for JSON parsers, make it a required dependency.
This fixes a common and easy-to-make error of forgetting to include it.

This does not preclude the use of websocket_client fork since users can
still override the dependency in their projects, but in that case, the
override will now be explicit.